Android 14.0.0 release 29

# -----BEGIN PGP SIGNATURE-----
#
# iF0EABECAB0WIQRDQNE1cO+UXoOBCWTorT+BmrEOeAUCZeZW2wAKCRDorT+BmrEO
# eNa4AJ4y3g9N+GSMWb4ATQL0HQnI8VYtlACffot1dTOjGYzZYfsSA0PgGq84Qws=
# =cwhq
# -----END PGP SIGNATURE-----
# gpg: Signature made Tue Mar  5 01:18:51 2024 EET
# gpg:                using DSA key 4340D13570EF945E83810964E8AD3F819AB10E78
# gpg: Good signature from "The Android Open Source Project <initial-contribution@android.com>" [marginal]
# gpg: initial-contribution@android.com: Verified 2347 signatures in the past
#      2 years.  Encrypted 4 messages in the past 2 years.
# gpg: WARNING: This key is not certified with sufficiently trusted signatures!
# gpg:          It is not certain that the signature belongs to the owner.
# Primary key fingerprint: 4340 D135 70EF 945E 8381  0964 E8AD 3F81 9AB1 0E78

# By Ankit Goyal (21) and others
# Via Automerger Merge Worker (159) and others
* tag 'android-14.0.0_r29': (47 commits)
  Revert^3 "gralloc4: Upgrade Allocator to AIDL2"
  Revert "Use memfd on placeholder buffers"
  Fix errors in Allocator::isSupported
  gralloc4: Allow GPU to read RAW12
  gralloc4: Warn about invalid usage (rather than error)
  Use memfd on placeholder buffers
  BufferAccess inputs should be width/height and not right/left
  BufferAccess inputs should be width/height and not right/left
  Gralloc4: Common Mapper code should not depend on hidl
  Revert "Add post-submit test for gralloc"
  Revert^2 "gralloc4: Upgrade Allocator to AIDL2"
  Preserve as many error codes as possible in map/unmap paths.
  Correct the error check on the lockBuffer path
  gralloc4: Add support for faeval-secure heap
  Allocate placeholder buffers for metadata-only allocations
  Revert "Support Dry allocation"
  Support Dry allocation
  libvendorgraphicbuffer: Use mapper::get for pixel metadata
  gralloc4: Use standard encode/decode
  gralloc4: Fix possible overflow in dmabuf_sanity_check()
  ...

+++ gralloc4/src/core/mali_gralloc_bufferdescriptor.h
@@ -51,7 +51,7 @@ struct buffer_descriptor_t
 	 * allocate the buffer.
 	 */
 	uint64_t alloc_sizes[MAX_PLANES];
-	int pixel_stride;
+	uint64_t pixel_stride;
 	uint64_t alloc_format;
 	uint32_t fd_count;
 	uint32_t plane_count;
@@ -86,11 +86,11 @@ struct buffer_descriptor_t
 			"name(%s)"
 			"reserved_size(%" PRIu64 ") "
 			"alloc_sizes(%" PRIu64 ", %" PRIu64 ", %" PRIu64 ")"
-			"pixel_stride(%d) alloc_format(0x%" PRIx64 ") fd_count(%d) "
+			"pixel_stride(%" PRIu64 ") alloc_format(0x%" PRIx64 ") fd_count(%d) "
 			"plane_count(%u) "
-			"plane[0](offset %" PRId64 ", idx %u, size %" PRIu64 " byte_stride %u, wh %u %u)"
-			"plane[1](offset %" PRId64 ", idx %u, size %" PRIu64 " byte_stride %u, wh %u %u)"
-			"plane[2](offset %" PRId64 ", idx %u, size %" PRIu64 " byte_stride %u, wh %u %u)"
+			"plane[0](offset %" PRId64 ", idx %u, size %" PRIu64 " byte_stride %" PRIu64 ", wh %" PRIu64 " %" PRIu64 ")"
+			"plane[1](offset %" PRId64 ", idx %u, size %" PRIu64 " byte_stride %" PRIu64 ", wh %" PRIu64 " %" PRIu64 ")"
+			"plane[2](offset %" PRId64 ", idx %u, size %" PRIu64 " byte_stride %" PRIu64 ", wh %" PRIu64 " %" PRIu64 ")"
 			"\n",
 			str.c_str(),
 			width, height,
