+++ gralloc4/src/mali_gralloc_formats.h
@@ -359,8 +359,8 @@ typedef struct mali_gralloc_format_caps mali_gralloc_format_caps;
 
 void mali_gralloc_adjust_dimensions(const uint64_t internal_format,
                                     const uint64_t usage,
-                                    int* const width,
-                                    int* const height);
+                                    uint64_t* const width,
+                                    uint64_t* const height);
 
 uint64_t mali_gralloc_select_format(const uint64_t req_format,
                                     const mali_gralloc_format_type type,
